made qcow build conditional, also introduced new_install_mode.go to refactor code to pass codefactor
