+++ scripts/package-harvester-os
@@ -71,10 +71,12 @@ rm ${ARTIFACTS_DIR}/${ISO_PREFIX}.iso.sha256
 xorriso -osirrox on -indev ${ARTIFACTS_DIR}/${ISO_PREFIX}.iso -extract /rootfs.squashfs ${ARTIFACTS_DIR}/${PROJECT_PREFIX}-rootfs-${ARCH}.squashfs
 
 # Build qcow image
-echo "generating harvester install mode qcow"
-qemu-img create -f qcow2 -o size=150G ${ARTIFACTS_DIR}/${PROJECT_PREFIX}-amd64.qcow2
-qemu-system-x86_64 -nographic -m 8192 -serial mon:stdio -serial file:harvester-installer.log -nic none -drive file=${ARTIFACTS_DIR}/${PROJECT_PREFIX}-amd64.qcow2,if=virtio,cache=writeback,discard=ignore,format=qcow2 -boot d -cdrom ${ARTIFACTS_DIR}/${PROJECT_PREFIX}-amd64.iso -kernel ${ARTIFACTS_DIR}/${PROJECT_PREFIX}-vmlinuz-amd64 -append "cdroot root=live:CDLABEL=COS_LIVE rd.live.dir=/ rd.live.squashimg=rootfs.squashfs console=ttyS1,115200 console=ttyS0,115200 rd.cos.disable net.ifnames=1 harvester.install.mode=install harvester.install.device=/dev/vda harvester.install.automatic=true harvester.install.powerOff=true harvester.os.password=password" -initrd ${ARTIFACTS_DIR}/${PROJECT_PREFIX}-initrd-amd64 -boot once=d
-tail -100 harvester-installer.log
+if [ "${BUILD_QCOW}" == "true" ]; then
+  echo "generating harvester install mode qcow"
+  qemu-img create -f qcow2 -o size=150G ${ARTIFACTS_DIR}/${PROJECT_PREFIX}-amd64.qcow2
+  qemu-system-x86_64 -nographic -m 8192 -serial mon:stdio -serial file:harvester-installer.log -nic none -drive file=${ARTIFACTS_DIR}/${PROJECT_PREFIX}-amd64.qcow2,if=virtio,cache=writeback,discard=ignore,format=qcow2 -boot d -cdrom ${ARTIFACTS_DIR}/${PROJECT_PREFIX}-amd64.iso -kernel ${ARTIFACTS_DIR}/${PROJECT_PREFIX}-vmlinuz-amd64 -append "cdroot root=live:CDLABEL=COS_LIVE rd.live.dir=/ rd.live.squashimg=rootfs.squashfs console=ttyS1,115200 console=ttyS0,115200 rd.cos.disable net.ifnames=1 harvester.install.mode=install harvester.install.device=/dev/vda harvester.install.automatic=true harvester.install.powerOff=true harvester.os.password=password" -initrd ${ARTIFACTS_DIR}/${PROJECT_PREFIX}-initrd-amd64 -boot once=d
+  tail -100 harvester-installer.log
+fi
 # Write checksum
 cd ${ARTIFACTS_DIR}
 CHECKSUM_FILE=${ISO_PREFIX}.sha512
